+++ app/routes/update.py
@@ -1,4 +1,4 @@
-import app.helpers.tasks as tasks
+import app.tasks as tasks
 from app.server import server
 from flask import request, abort
 from os import path, getcwd
@@ -10,7 +10,7 @@ def update():
     commit = request.form['commit']
     with open(path.join(getcwd(), "deploy.txt")) as f:
         if commit == f.read():
-            tasks.update.delay(commit)
+            tasks.update.update.delay(commit)
             return 'success', 200
         else:
             return abort(403)
