+++ misc/__init__.py
@@ -1,0 +1,3 @@
+from os import path
+
+md_exe = path.join(path.dirname(__file__), 'markdown.js')
